+++ tests/HTMLConverterTest.php
@@ -19,7 +19,7 @@ use PHPUnit\Framework\TestCase;
 
 /**
  * @group converter
- * @coversDefaultClass League\Csv\HTMLConverter
+ * @coversDefaultClass \League\Csv\HTMLConverter
  */
 class HTMLConverterTest extends TestCase
 {
@@ -60,13 +60,11 @@ class HTMLConverterTest extends TestCase
     }
 
     /**
-     * @covers ::__construct
-     * @covers ::table
-     * @covers ::tr
-     * @covers ::td
      * @covers ::convert
+     * @covers ::appendTableHeader
+     * @covers ::styleTableElement
      */
-    public function testToHTMLWithHeaders()
+    public function testToHTMLWithTHeadTableSection()
     {
         $csv = Reader::createFromPath(__DIR__.'/data/prenoms.csv', 'r')
             ->setDelimiter(';')
@@ -86,21 +84,81 @@ class HTMLConverterTest extends TestCase
             ->tr('data-record-offset')
         ;
 
-        $html = $converter->convert($records, $records->getHeader());
+        $headers = $records->getHeader();
+
+        $html = $converter->convert($records, $headers);
         self::assertContains('<table class="table-csv-data" id="test">', $html);
         self::assertContains('<th scope="col">prenoms</th>', $html);
         self::assertContains('<thead>', $html);
         self::assertContains('<tbody>', $html);
         self::assertNotContains('<tfoot>', $html);
+        self::assertNotContains('<thead><tr data-record-offset="0"></tr></thead>', $html);
+    }
 
-        $html = $converter->convert($records, [], $records->getHeader());
+    /**
+     * @covers ::convert
+     * @covers ::appendTableHeader
+     * @covers ::styleTableElement
+     */
+    public function testToHTMLWithTFootTableSection()
+    {
+        $csv = Reader::createFromPath(__DIR__.'/data/prenoms.csv', 'r')
+            ->setDelimiter(';')
+            ->setHeaderOffset(0)
+        ;
+
+        $stmt = (new Statement())
+            ->offset(3)
+            ->limit(5)
+        ;
+
+        $records = $stmt->process($csv);
+
+        $converter = (new HTMLConverter())
+            ->table('table-csv-data', 'test')
+            ->td('title')
+            ->tr('data-record-offset')
+        ;
+
+        $headers = $records->getHeader();
+
+        $html = $converter->convert($records, [], $headers);
         self::assertContains('<table class="table-csv-data" id="test">', $html);
         self::assertContains('<th scope="col">prenoms</th>', $html);
         self::assertNotContains('<thead>', $html);
         self::assertContains('<tbody>', $html);
         self::assertContains('<tfoot>', $html);
+        self::assertNotContains('<tfoot><tr data-record-offset="0"></tr></tfoot>', $html);
+    }
 
-        $html = $converter->convert($records, $records->getHeader(), $records->getHeader());
+    /**
+     * @covers ::convert
+     * @covers ::appendTableHeader
+     * @covers ::styleTableElement
+     */
+    public function testToHTMLWithBothTableHeaderSection()
+    {
+        $csv = Reader::createFromPath(__DIR__.'/data/prenoms.csv', 'r')
+            ->setDelimiter(';')
+            ->setHeaderOffset(0)
+        ;
+
+        $stmt = (new Statement())
+            ->offset(3)
+            ->limit(5)
+        ;
+
+        $records = $stmt->process($csv);
+
+        $converter = (new HTMLConverter())
+            ->table('table-csv-data', 'test')
+            ->td('title')
+            ->tr('data-record-offset')
+        ;
+
+        $headers = $records->getHeader();
+
+        $html = $converter->convert($records, $headers, $headers);
         self::assertContains('<table class="table-csv-data" id="test">', $html);
         self::assertContains('<thead>', $html);
         self::assertContains('<tbody>', $html);